[SearchBundle] Allow easier switching between word and nGram analyzers
